+++ src/tools/dalle/Render/index.tsx
@@ -1,6 +1,6 @@
 import { ActionIcon, ImageGallery } from '@lobehub/ui';
 import { Download } from 'lucide-react';
-import { memo, useState } from 'react';
+import { memo, useRef } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
 import GalleyGrid from '@/components/GalleyGrid';
@@ -11,17 +11,17 @@ import { DallEImageItem } from '@/types/tool/dalle';
 import ImageItem from './Item';
 
 const DallE = memo<BuiltinRenderProps<DallEImageItem[]>>(({ content, messageId }) => {
-  const [current, setCurrent] = useState<number>(0);
+  const currentRef = useRef(0);
 
   const handleDownload = async () => {
     // 1. Retrieve the blob URL of an image by its imageId
-    const id = content[current]?.imageId;
+    const id = content[currentRef.current]?.imageId;
     if (!id) return;
-    const { url } = await fileService.getFile(id);
+    const { url, name } = await fileService.getFile(id);
     // 2. Download the image
     const link = document.createElement('a');
     link.href = url;
-    link.download = 'image.png'; // 设置下载的文件名
+    link.download = name; // 设置下载的文件名
     link.click();
   };
 
@@ -33,13 +33,11 @@ const DallE = memo<BuiltinRenderProps<DallEImageItem[]>>(({ content, messageId }
         preview={{
           // 切换图片时设置
           onChange: (current: number) => {
-            setCurrent(current);
+            currentRef.current = current;
           },
           // 点击预览显示时设置
           onVisibleChange: (visible: boolean, _prevVisible: boolean, current: number) => {
-            if (visible) {
-              setCurrent(current);
-            }
+            currentRef.current = current;
           },
           toolbarAddon: <ActionIcon color={'#fff'} icon={Download} onClick={handleDownload} />,
         }}